// web/src/lib/api-types.ts
export interface components {
  schemas: {
// ...
    SkillOut: {
      /** Name */
      name: string;
      /**
       * Category
       * @enum {string}
       */
      category: "frontend" | "backend" | "devops" | "tools" | "soft";
      /** Level */
      level?: number | null;
      /** Icon */
      icon?: string | null;
      /**
       * Display Order
       * @default 0
       */
      display_order: number;
      /**
       * Is Featured
       * @default false
       */
      is_featured: boolean;
      /** Id */
      id: number;
      /**
       * Created At
       * Format: date-time
       */
      created_at: string;
      /**
       * Updated At
       * Format: date-time
       */
      updated_at: string;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}